planets to long, counter to double

+++ src/wfpt.pyx
@@ -203,7 +203,7 @@ def wiener_like_rlddm_2step(np.ndarray[double, ndim=1] x1, # 1st-stage RT
     # parameters added for two-step
 
 
-    cdef np.ndarray[int, ndim=1] planets
+    cdef np.ndarray[long, ndim=1] planets
     cdef np.ndarray[double, ndim=1] counter = np.zeros(comb(nstates,2,exact=True))
     # cdef np.ndarray[double, ndim=1] Qmb
     # cdef double dtq
